Move page structures to toplevel.
